+++ data-manager/api/src/test/java/org/orbisgis/orbisdata/datamanager/api/dataset/IJdbcTableTest.java
@@ -277,6 +277,9 @@ public class IJdbcTableTest {
         }
 
         @Override
+        public Object get(int column) {return null;}
+
+        @Override
         public boolean isLinked() {
             return false;
         }
@@ -514,6 +517,11 @@ public class IJdbcTableTest {
 
         @Override
         public BigDecimal getBigDecimal(int i) {
+            return null;
+        }
+
+        @Override
+        public Object get(String column) throws Exception {
             return null;
         }