+++ docs/CHANGELOG.md
@@ -17,3 +17,4 @@ make it compatible with `smile` API)
 + Fix bug on DSL built tables.
 + Add a method `getSummary()` on the `IDataSet` interface.
 + Add the `ProgressMonitor` mechanism.
++ Add get*Type*() methods on `ITable` interface